reorganized AC namespace
